atelier3/requirements.txt
apache-airflow==2.9.0
pandas==1.5.3
pytest==8.4.0
python-dotenv==1.1.0
apache-airflow-providers-postgres==6.1.0
requests==2.32.4
duckdb==0.8.1